LINQ Start
L: Language
IN: Intergrated
Q: Query

Entity - Definition of a complete database entity (sql table)

POCO - Plain Ordinary Common Object
- flat data set which contains primitive data types such as int, decimal, bool, etc...

POCOs DO NOT have structure (object, List, array)
Example of a POCO: What do you really need for a drop down list? you need 2 pieces of data, a value text and a disaply text. it might not be the entire entity in a POCO. if you want the album info you can get it but you might not want the release year and release label, maybe you just want the artist id and track name. It is a subset of data

DTO - Data Tranfer Object
- Contrains a structure. which means you can have an object in the data set. you can have a list of <t> where a POCO would not have a structure. DTO can have a list, array, or even a POCO.

Because we are adding DTOs and POCOs we are going to add two new folders and we will add them in on friday most likely.

To Open LINQ:
Open start menu -> search for LINQPad4 open it up. Linqpad is free you are going to want to download it at home.

What is linqpad - it is a scratchpad sort of thing, allowing you to test queries before using them in your application.

First thing you will need is a connection to a database. It can be found on the notes under linq intro.

Adding a connection in LINQ: click the add connection button in the top left
on the new box that opens you pick the context you want (LINQ to SQL) when you move it to your project it will be linq to entities.
Click Next
under the connection page put a . for the server go to new or existing database and choose the database you want then you will be connected (in this case it will be Chinook) leave all the other defualys.

When starting to use linq dont forget to set the connection you want to use in the drop down menu on the top of the linqpad.

In the database viewer you can see the properties and navigational properties under each table.

To Open a new tab in Linq: Click the plus next to the queary header at the top of LinqPad

Execute query is the green arrow
Stop Query is the red box

C# Statments is almost exactly how it will look in your application when you add it in. You can do this mode by selecting it in the language drop down menu on the top of LinqPad

First Linq Statement
Printing the albums by typing Albums and running it. This will list all the albums in the albums table.

Ctrl + R hides/shows results

by clicking the SQL tag at the top of the results it will convert your statment to an SQL Query

Clicking the Lambda symbol (upsidedown y) shows the Linq Query version.

Second Linq Query
from rowPlaceholder in Albums
select rowPlaceholder

This should give you the same result as the first query.
rowPlaceholder represents a row.

The syntax we are using right now is call query syntax.

//will work the same as C# to give you comments in Linq

There is another type of syntax called method syntax which looks like what you find under the lambda symbol verion of linq

By highlighting the query you want to run you can run only that selection just like in SQL

Next Linq Query
//query syntax
from rowPlaceholder in Albums
orderby rowPlaceholder.ReleaseYear
select rowPlaceholder
//method syntax
Albums
   .Select (rowPlaceholder => rowPlaceholder)

Where Clause
//query syntax
from rowPlaceholder in Albums
where rowPlaceholder.ArtistId.Equals("22")
orderby rowPlaceholder.ReleaseYear
select rowPlaceholder
//method syntax
Albums
   .Where (rowPlaceholder => rowPlaceholder.ArtistId.Equals ("22"))
   .OrderBy (rowPlaceholder => rowPlaceholder.ReleaseYear)